+++ test/linkify_test.dart
@@ -33,7 +33,14 @@ void main() {
   test('Parses only link', () {
     expectListEqual(
       linkify("https://example.com"),
-      [LinkElement("https://example.com")],
+      [UrlElement("https://example.com", "example.com")],
+    );
+  });
+
+  test('Parses only link with no humanize', () {
+    expectListEqual(
+      linkify("https://example.com", options: LinkifyOptions(humanize: false)),
+      [UrlElement("https://example.com")],
     );
   });
 
@@ -41,9 +48,9 @@ void main() {
     expectListEqual(
       linkify("https://example.com https://google.com"),
       [
-        LinkElement("https://example.com"),
+        UrlElement("https://example.com", "example.com"),
         TextElement(" "),
-        LinkElement("https://google.com"),
+        UrlElement("https://google.com", "google.com"),
       ],
     );
   });
@@ -55,9 +62,24 @@ void main() {
       ),
       [
         TextElement("Lorem ipsum dolor sit amet "),
-        LinkElement("https://example.com"),
+        UrlElement("https://example.com", "example.com"),
         TextElement(" "),
-        LinkElement("https://google.com"),
+        UrlElement("https://google.com", "google.com"),
+      ],
+    );
+  });
+
+  test('Parses links with text with no humanize', () {
+    expectListEqual(
+      linkify(
+        "Lorem ipsum dolor sit amet https://example.com https://google.com",
+        options: LinkifyOptions(humanize: false),
+      ),
+      [
+        TextElement("Lorem ipsum dolor sit amet "),
+        UrlElement("https://example.com"),
+        TextElement(" "),
+        UrlElement("https://google.com"),
       ],
     );
   });
@@ -68,9 +90,9 @@ void main() {
         "https://google.com\nLorem ipsum\ndolor sit amet\nhttps://example.com",
       ),
       [
-        LinkElement("https://google.com"),
+        UrlElement("https://google.com", "google.com"),
         TextElement("\nLorem ipsum\ndolor sit amet\n"),
-        LinkElement("https://example.com"),
+        UrlElement("https://example.com", "example.com"),
       ],
     );
   });
@@ -79,6 +101,26 @@ void main() {
     expectListEqual(
       linkify("person@example.com"),
       [EmailElement("person@example.com")],
+    );
+  });
+
+  test('Parses email and link', () {
+    expectListEqual(
+      linkify("person@example.com at https://google.com"),
+      [
+        EmailElement("person@example.com"),
+        TextElement(" at "),
+        UrlElement("https://google.com", "google.com")
+      ],
+    );
+  });
+
+  test("Doesn't parses email and link with no linkifiers", () {
+    expectListEqual(
+      linkify("person@example.com at https://google.com", linkifiers: []),
+      [
+        TextElement("person@example.com at https://google.com"),
+      ],
     );
   });
 }