+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+## [2.0.0] - 2019-12-27
+
+- Change `LinkTypes` to `Linkifier`
+  - Supports custom linkifiers
+- Change `LinkElement` to `UrlElement` to better reflect `UrlLinkifier` (link != URL)
+- Change `humanize` option to `LinkifyOptions`
+- Enabled `humanize` by default
+
 ## [1.0.1] - 2019-03-23
 
-- Republish to fix maintenance score.
+- Republish to fix maintenance score
 
 ## [1.0.0] - 2019-03-23